Change EventPopup to use the generic Popup component
