+++ evento/src/views/EventPopup/components/AttendeeList/AttendeeList.scss
@@ -1,0 +1,32 @@
+.AttendeeList {
+	.header {
+		font-size: 16px;
+		font-weight: 500;
+		opacity: 0.9;
+		margin-bottom: 4px;
+		margin-top: 12px;
+
+		.attendee-count {
+			font-size: 12px;
+			font-weight: 400;
+			position: relative;
+			bottom: 1px;
+		}
+	}
+
+	.attendee-container {
+		padding-left: 4px;
+		margin: 4px;
+		height: 124px !important;
+
+		li {
+			list-style-type: none;
+
+			p {
+				 margin-top: 2px;
+				 margin-bottom: 2px;
+				 font-size: 14px;
+			}
+		}
+	}
+}
