+++ evento/src/views/EventPopup/components/Map/Map.scss
@@ -1,0 +1,5 @@
+.Map, .Map img {
+	display: block;
+	margin: auto;
+	cursor: pointer;
+}
